add newPassWindow binding functionality

+++ VaccinationPassportLibrary/Models/Vaccine.cs
@@ -11,6 +11,6 @@ namespace VaccinationPassportLibrary.Models
         public int Id { get; set; } 
         public Disease Disease { get; set; }
         public string VaccineName { get; set; }
-        public float Dose { get; set; }
+        public double Dose { get; set; }
     }
 }
